BF-90 deisgn with real data and tested

+++ frontend/lib/models/user_profile.dart
@@ -5,9 +5,9 @@ class UserProfile {
 
   final String email;
 
-  final Double height;
+  final double height;
 
-  final Double weight;
+  final double weight;
 
   final String birthdate;
 
@@ -22,8 +22,8 @@ class UserProfile {
     return UserProfile(
         username: json['username'] as String,
         email: json['email'] as String,
-        height: json['height'] as Double,
-        weight: json['weight'] as Double,
+        height: json['height'],
+        weight: json['weight'],
         birthdate: json['birthdate'] as String);
   }
